moves mocks into separate folder
