+++ spec/file_store_mocks/mock_file.rb
