+++ spec/file_store_mocks/mock_file_system.rb
