+++ spec/file_store_mocks/mock_store.rb
@@ -10,6 +10,5 @@ class MockStore
   end
 
   def read_from_file
-
   end
 end
